requirements.txt
# Core deep learning
torch==2.2.0+cu118
torchvision==0.17.0+cu118
torchaudio==2.2.0+cu118
timm==0.9.2

# Object detection
ultralytics==8.3.70

# Medical imaging
pydicom==3.0.1

# Image processing
opencv-python==4.11.0.86
scikit-image==0.23.2
albumentations==1.4.20
pillow==10.3.0

# Scientific computing
numpy==1.26.4
scipy==1.11.4

# Visualisation
matplotlib==3.9.0
PyQt5==5.15.9
